refactor: moving triggers package
