+++ src/scripts/triggers/detectPageLoad.tab.js
