+++ src/scripts/triggers/notify.bg.js
